Rename the new DB tables to use their old names: Props, Entities> Moved the new funcs that replaced the views into the appropriate entity/resource files. More checks for xrlint More dev docs
