+++ registry/registry.go
@@ -218,7 +218,7 @@ func NewRegistry(tx *Tx, id string, regOpts ...RegOpt) (*Registry, *XRError) {
 
 	reg.Self = reg
 	reg.Entity.Registry = reg
-	reg.FullEntityInsert()
+	reg.EntityInsert()
 	reg.Capabilities = DefaultCapabilities.Clone()
 	reg.Model = &Model{
 		Registry: reg,
@@ -350,8 +350,8 @@ func FindRegistryBySID(tx *Tx, sid string, accessMode int) (*Registry, *XRError)
 	tx.Registry = reg
 	tx.AddRegistry(reg)
 
-	// UsesXref lives on the raw Registries table (not FullEntities/
-	// FullTreeTable, since it's a plain internal flag, not a real
+	// UsesXref lives on the raw Registries table (not Entities/
+	// Props, since it's a plain internal flag, not a real
 	// attribute), so it needs its own tiny supplemental lookup here -
 	// a single indexed PK read, once per Tx.
 	results := Query(tx, `SELECT UsesXref FROM Registries WHERE SID=?`, sid)
@@ -816,7 +816,7 @@ func (reg *Registry) UpsertGroupWithObject(gType string, id string, obj Object, 
 			gm.SID, g.Path, g.Abstract,
 			g.Plural, g.Singular)
 
-		g.FullEntityInsert()
+		g.EntityInsert()
 
 		// Use the ID passed as an arg, not from the metadata, as the true
 		// ID. If the one in the metadata differs we'll flag it down below
@@ -1019,7 +1019,7 @@ func GenerateQuery(reg *Registry, what string, paths []string, filters [][]*Filt
 `
 
 		sortJoin = `
-  LEFT JOIN FullTreeTable AS sj ON (
+  LEFT JOIN Props AS sj ON (
     sj.RegSID = ft.RegSID AND
     sj.Path = substring_index(ft.Path, '/', ` + slashCount + `) AND
     sj.PropName = '` + sortKey + `')
@@ -1030,7 +1030,7 @@ func GenerateQuery(reg *Registry, what string, paths []string, filters [][]*Filt
 	query = `
 SELECT
   ft.RegSID,ft.Type,ft.Plural,ft.Singular,ft.ParentSID,ft.eSID,ft.UID,ft.Abstract,ft.Path,ft.PropName,ft.PropValue,ft.PropType,ft.IsSystemProp
-  FROM FullTreeTable AS ft` + sortJoin + `
+  FROM Props AS ft` + sortJoin + `
   WHERE ft.RegSID=?
 `
 
@@ -1065,7 +1065,7 @@ ft.eSID IN ( -- eSID from query
   -- This "RECURSIVE" stuff finds all parents
   WITH RECURSIVE cte(eSID,Type,ParentSID,Path) AS (
     -- This defines the init set of rows of the query. We'll recurse later on
-    SELECT eSID,Type,ParentSID,Path FROM FullEntities
+    SELECT eSID,Type,ParentSID,Path FROM Entities
     WHERE eSID in ( -- start of the OR Filter groupings`
 		// This section will find all matching entities
 		firstOr := true
@@ -1079,7 +1079,7 @@ ft.eSID IN ( -- eSID from query
       -- start of one Filter AND grouping (expr1 AND expr2).
       -- Find all SIDs for the leaves for entities (SIDs) of interest.
       SELECT list.eSID FROM (
-        SELECT count(*) as cnt,e2.eSID,e2.Path FROM FullEntities AS e1
+        SELECT count(*) as cnt,e2.eSID,e2.Path FROM Entities AS e1
         RIGHT JOIN (
           -- start of expr1 - below finds SearchNodes/SIDs of interest`
 			firstAnd := true
@@ -1143,7 +1143,7 @@ ft.eSID IN ( -- eSID from query
 					// 2+ rows at matching more than one filter expression
 					check += " GROUP BY eSID" // " LIMIT 1"
 					query += `
-          (SELECT eSID,Type,Path FROM FullTreeTable  -- FILTER_PRESENT
+          (SELECT eSID,Type,Path FROM Props  -- FILTER_PRESENT
            WHERE RegSID=? AND ` + check + ")" // Need () for groupBy/limit
 
 				} else if filter.Operator == FILTER_ABSENT { // ?filter=xxx=null
@@ -1154,9 +1154,9 @@ ft.eSID IN ( -- eSID from query
 					// BINARY means case-sensitive for that operand
 					query += `
           -- Entities that don't have the specified prop
-          SELECT e.eSID,e.Type,e.Path FROM FullEntities AS e
+          SELECT e.eSID,e.Type,e.Path FROM Entities AS e
           WHERE e.RegSID=? AND e.Abstract=? AND
-            NOT EXISTS (SELECT 1 FROM FullTreeTable WHERE
+            NOT EXISTS (SELECT 1 FROM Props WHERE
               RegSID=e.RegSID AND eSID=e.eSID AND (` + binary + ` ` +
 						propNameSearch + `))`
 
@@ -1183,7 +1183,7 @@ ft.eSID IN ( -- eSID from query
 					}
 					check += ")"
 					query += `
-          SELECT eSID,Type,Path FROM FullTreeTable
+          SELECT eSID,Type,Path FROM Props
             WHERE RegSID=? AND ` + check
 
 				} else if filter.Operator == FILTER_NOT_EQUAL { // ?filter=x!=z
@@ -1192,9 +1192,9 @@ ft.eSID IN ( -- eSID from query
 					// BINARY means case-sensitive for that operand
 					query += `
           -- Entities that don't have the specified prop
-          SELECT e.eSID,e.Type,e.Path FROM FullEntities AS e
+          SELECT e.eSID,e.Type,e.Path FROM Entities AS e
           WHERE e.RegSID=? AND e.Abstract=? AND
-            NOT EXISTS (SELECT 1 FROM FullTreeTable WHERE
+            NOT EXISTS (SELECT 1 FROM Props WHERE
               RegSID=e.RegSID AND eSID=e.eSID AND (` + binary + ` ` +
 						propNameSearch + ` AND `
 
@@ -1242,7 +1242,7 @@ ft.eSID IN ( -- eSID from query
 						" THEN CAST(PropValue AS DECIMAL) " + sqlOp + " CAST(? AS DECIMAL)" +
 						" ELSE PropValue " + FILTER_CI_COLLATE + " " + sqlOp + " ? END))"
 					query += `
-          SELECT eSID,Type,Path FROM FullTreeTable
+          SELECT eSID,Type,Path FROM Props
             WHERE RegSID=? AND ` + check
 
 				} else {
@@ -1257,7 +1257,7 @@ ft.eSID IN ( -- eSID from query
         -- expressions in each filter (the ANDs), are branches to return.
         -- Note we return the Path of each Leaf, not the path of the matching
         -- entity. The entity that matches isn't important.
-        JOIN FullEntities AS e2 ON (
+        JOIN Entities AS e2 ON (
           (
             (
               -- Non-meta objects, just compare the Path
@@ -1292,7 +1292,7 @@ ft.eSID IN ( -- eSID from query
     -- entities, up to root of Reg.
     UNION DISTINCT SELECT
       e.eSID,e.Type,e.ParentSID,e.Path
-    FROM FullEntities AS e
+    FROM Entities AS e
     INNER JOIN cte ON
       (
         -- Find its parent
