+++ tests/format_test.go
@@ -1941,3 +1941,1692 @@ func TestFormatCompatModes(t *testing.T) {
 `)
 
 }
+
+// TestXrefFormatCascadeOnDirectSave (gaps item 1, remaining half) checks
+// that formatvalidated correctly cascades into an xref source through the
+// NORMAL per-Resource save path (runCascade() called directly from
+// ValidateResource(), not via Registry.VerifyData()'s model-driven
+// revalidation pass, which is what tests/xref_model_revalidation_test.go
+// covers instead). format_test.go has zero xref mentions, so this was
+// completely unverified before.
+func TestXrefFormatCascadeOnDirectSave(t *testing.T) {
+	reg := NewRegistry("TestXrefFormatCascadeOnDirectSave")
+	defer PassDeleteReg(t, reg)
+
+	model := registry.Model{}
+	gm, _ := model.AddGroupModel("dirs", "dir")
+	rm, _ := gm.AddResourceModel("files", "file", 0, true, true)
+	rm.SetValidateFormat(true)
+
+	XHTTP(t, reg, "PUT", "/modelsource", model.MustUserMarshal("", "  "),
+		200, `{
+  "groups": {
+    "dirs": {
+      "plural": "dirs",
+      "singular": "dir",
+      "resources": {
+        "files": {
+          "plural": "files",
+          "singular": "file",
+          "maxversions": 0,
+          "setversionid": true,
+          "hasdocument": true,
+          "versionmode": "manual",
+          "singleversionroot": false,
+          "validateformat": true,
+          "validatecompatibility": false,
+          "strictvalidation": false
+        }
+      }
+    }
+  }
+}
+`)
+
+	// Create the target with a valid "numbers" document straight away.
+	XHTTP(t, reg, "PUT", "/dirs/d1/files/f1$details", `{
+  "format": "numbers",
+  "file": "1"
+}`, 201, `{
+  "fileid": "f1",
+  "versionid": "1",
+  "self": "http://localhost:8181/dirs/d1/files/f1$details",
+  "xid": "/dirs/d1/files/f1",
+  "epoch": 1,
+  "isdefault": true,
+  "createdat": "2026-07-27T00:24:29.606428854Z",
+  "modifiedat": "2026-07-27T00:24:29.606428854Z",
+  "ancestorid": "1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/f1/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/f1/versions",
+  "versionscount": 1
+}
+`)
+
+	// Create the xref source pointing at it.
+	XHTTP(t, reg, "PUT", "/dirs/d1/files/fx/meta",
+		`{"xref":"/dirs/d1/files/f1"}`, 201, `{
+  "fileid": "fx",
+  "self": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "xid": "/dirs/d1/files/fx/meta",
+  "xref": "/dirs/d1/files/f1",
+  "epoch": 1,
+  "createdat": "2026-07-27T00:24:31.661658328Z",
+  "modifiedat": "2026-07-27T00:24:31.661658328Z",
+  "readonly": false,
+
+  "defaultversionid": "1",
+  "defaultversionurl": "http://localhost:8181/dirs/d1/files/fx/versions/1$details",
+  "defaultversionsticky": false
+}
+`)
+
+	// The mirror must already show formatvalidated=true.
+	XHTTP(t, reg, "GET", "/dirs/d1/files/fx$details", ``, 200,
+		`{
+  "fileid": "fx",
+  "versionid": "1",
+  "self": "http://localhost:8181/dirs/d1/files/fx$details",
+  "xid": "/dirs/d1/files/fx",
+  "epoch": 1,
+  "isdefault": true,
+  "createdat": "2026-07-27T00:24:33.708753687Z",
+  "modifiedat": "2026-07-27T00:24:33.708753687Z",
+  "ancestorid": "1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/fx/versions",
+  "versionscount": 1
+}
+`)
+
+	// Now directly PUT a new (still valid) document onto the TARGET - a
+	// perfectly normal save, no model change involved at all.
+	XHTTP(t, reg, "POST", "/dirs/d1/files/f1/versions", `{
+  "v2": {
+    "format": "numbers",
+    "file": "2"
+  }
+}`, 200, `{
+  "v2": {
+    "fileid": "f1",
+    "versionid": "v2",
+    "self": "http://localhost:8181/dirs/d1/files/f1/versions/v2$details",
+    "xid": "/dirs/d1/files/f1/versions/v2",
+    "epoch": 1,
+    "isdefault": true,
+    "createdat": "2026-07-27T00:24:35.911759653Z",
+    "modifiedat": "2026-07-27T00:24:35.911759653Z",
+    "ancestorid": "1",
+    "contenttype": "application/json",
+    "format": "numbers",
+    "formatvalidated": true
+  }
+}
+`)
+
+	// The xref mirror must reflect both the new default Version and the
+	// (still true) formatvalidated flag for it.
+	XHTTP(t, reg, "GET", "/dirs/d1/files/fx$details", ``, 200,
+		`{
+  "fileid": "fx",
+  "versionid": "v2",
+  "self": "http://localhost:8181/dirs/d1/files/fx$details",
+  "xid": "/dirs/d1/files/fx",
+  "epoch": 1,
+  "isdefault": true,
+  "createdat": "2026-07-27T00:24:37.975662554Z",
+  "modifiedat": "2026-07-27T00:24:37.975662554Z",
+  "ancestorid": "1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/fx/versions",
+  "versionscount": 2
+}
+`)
+}
+
+// TestXrefCompatModeChangeCascade (gaps item 5) combines a Version-content-
+// unrelated compatibility MODE change (meta.compatibility, an
+// instance-level attribute) with an active xref, exercising the normal
+// runCascade() fan-out path (as opposed to the model-driven
+// Registry.VerifyData() path already covered by
+// tests/xref_model_revalidation_test.go).
+func TestXrefCompatModeChangeCascade(t *testing.T) {
+	reg := NewRegistry("TestXrefCompatModeChangeCascade")
+	defer PassDeleteReg(t, reg)
+
+	model := registry.Model{}
+	gm, _ := model.AddGroupModel("dirs", "dir")
+	rm, _ := gm.AddResourceModel("files", "file", 0, true, true)
+	rm.SetValidateFormat(true)
+	rm.SetValidateCompatibility(true)
+
+	XHTTP(t, reg, "PUT", "/modelsource", model.MustUserMarshal("", "  "),
+		200, `{
+  "groups": {
+    "dirs": {
+      "plural": "dirs",
+      "singular": "dir",
+      "resources": {
+        "files": {
+          "plural": "files",
+          "singular": "file",
+          "maxversions": 0,
+          "setversionid": true,
+          "hasdocument": true,
+          "versionmode": "manual",
+          "singleversionroot": false,
+          "validateformat": true,
+          "validatecompatibility": true,
+          "strictvalidation": false
+        }
+      }
+    }
+  }
+}
+`)
+
+	// v1 sum=2, v2 sum=2 - equal, so compatible under any mode, including
+	// bidirectionally (needed since "full" checks both directions).
+	XHTTP(t, reg, "PUT", "/dirs/d1/files/f1$details", `{
+  "meta": { "compatibility": "backward" },
+  "versions": {
+    "v1": { "format": "numbers", "file": "2" },
+    "v2": { "format": "numbers", "file": "2" }
+  }
+}`, 201, `{
+  "fileid": "f1",
+  "versionid": "v2",
+  "self": "http://localhost:8181/dirs/d1/files/f1$details",
+  "xid": "/dirs/d1/files/f1",
+  "epoch": 1,
+  "isdefault": true,
+  "createdat": "2026-07-27T00:25:31.864526553Z",
+  "modifiedat": "2026-07-27T00:25:31.864526553Z",
+  "ancestorid": "v1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+  "compatibilityvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/f1/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/f1/versions",
+  "versionscount": 2
+}
+`)
+
+	XHTTP(t, reg, "PUT", "/dirs/d1/files/fx/meta",
+		`{"xref":"/dirs/d1/files/f1"}`, 201, `{
+  "fileid": "fx",
+  "self": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "xid": "/dirs/d1/files/fx/meta",
+  "xref": "/dirs/d1/files/f1",
+  "epoch": 1,
+  "createdat": "2026-07-27T00:25:33.901048058Z",
+  "modifiedat": "2026-07-27T00:25:33.901048058Z",
+  "readonly": false,
+  "compatibility": "backward",
+
+  "defaultversionid": "v2",
+  "defaultversionurl": "http://localhost:8181/dirs/d1/files/fx/versions/v2$details",
+  "defaultversionsticky": false
+}
+`)
+
+	XHTTP(t, reg, "GET", "/dirs/d1/files/fx/meta", ``, 200,
+		`{
+  "fileid": "fx",
+  "self": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "xid": "/dirs/d1/files/fx/meta",
+  "xref": "/dirs/d1/files/f1",
+  "epoch": 1,
+  "createdat": "2026-07-27T00:25:35.886918015Z",
+  "modifiedat": "2026-07-27T00:25:35.886918015Z",
+  "readonly": false,
+  "compatibility": "backward",
+
+  "defaultversionid": "v2",
+  "defaultversionurl": "http://localhost:8181/dirs/d1/files/fx/versions/v2$details",
+  "defaultversionsticky": false
+}
+`)
+	XHTTP(t, reg, "GET", "/dirs/d1/files/fx$details", ``, 200,
+		`{
+  "fileid": "fx",
+  "versionid": "v2",
+  "self": "http://localhost:8181/dirs/d1/files/fx$details",
+  "xid": "/dirs/d1/files/fx",
+  "epoch": 1,
+  "isdefault": true,
+  "createdat": "2026-07-27T00:25:37.948014926Z",
+  "modifiedat": "2026-07-27T00:25:37.948014926Z",
+  "ancestorid": "v1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+  "compatibilityvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/fx/versions",
+  "versionscount": 2
+}
+`)
+
+	// Change ONLY the compatibility MODE on the target - no version
+	// content change at all - equal sums remain compatible under "full"
+	// (bidirectional) too.
+	XHTTP(t, reg, "PATCH", "/dirs/d1/files/f1$details",
+		`{"meta":{"compatibility":"full"}}`, 200,
+		`{
+  "fileid": "f1",
+  "versionid": "v2",
+  "self": "http://localhost:8181/dirs/d1/files/f1$details",
+  "xid": "/dirs/d1/files/f1",
+  "epoch": 2,
+  "isdefault": true,
+  "createdat": "2026-07-27T00:25:39.976131708Z",
+  "modifiedat": "2026-07-27T00:25:40.114064757Z",
+  "ancestorid": "v1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+  "compatibilityvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/f1/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/f1/versions",
+  "versionscount": 2
+}
+`)
+
+	// The xref mirror must reflect BOTH the new mode and the refreshed
+	// compatibilityvalidated result.
+	XHTTP(t, reg, "GET", "/dirs/d1/files/fx/meta", ``, 200,
+		`{
+  "fileid": "fx",
+  "self": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "xid": "/dirs/d1/files/fx/meta",
+  "xref": "/dirs/d1/files/f1",
+  "epoch": 2,
+  "createdat": "2026-07-27T00:25:42.126789781Z",
+  "modifiedat": "2026-07-27T00:25:42.272503534Z",
+  "readonly": false,
+  "compatibility": "full",
+
+  "defaultversionid": "v2",
+  "defaultversionurl": "http://localhost:8181/dirs/d1/files/fx/versions/v2$details",
+  "defaultversionsticky": false
+}
+`)
+	XHTTP(t, reg, "GET", "/dirs/d1/files/fx$details", ``, 200,
+		`{
+  "fileid": "fx",
+  "versionid": "v2",
+  "self": "http://localhost:8181/dirs/d1/files/fx$details",
+  "xid": "/dirs/d1/files/fx",
+  "epoch": 2,
+  "isdefault": true,
+  "createdat": "2026-07-27T00:25:44.2221894Z",
+  "modifiedat": "2026-07-27T00:25:44.356102502Z",
+  "ancestorid": "v1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+  "compatibilityvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/fx/versions",
+  "versionscount": 2
+}
+`)
+}
+
+// TestXrefFormatValidationFailureCascade verifies that a FAILED format
+// validation state (formatvalidated=false/compatibilityvalidated=false,
+// via an "unknown format" document - the only kind of format failure
+// that doesn't outright reject the save, see EnsureCompat()) on the xref
+// TARGET is correctly mirrored into the xref SOURCE too, not just the
+// "everything is valid" happy path already covered by
+// TestXrefFormatCascadeOnDirectSave. This is created BEFORE the xref
+// exists, so the xref creation itself must pick up the already-failed
+// state.
+func TestXrefFormatValidationFailureCascade(t *testing.T) {
+	reg := NewRegistry("TestXrefFormatValidationFailureCascade")
+	defer PassDeleteReg(t, reg)
+
+	model := registry.Model{}
+	gm, _ := model.AddGroupModel("dirs", "dir")
+	rm, _ := gm.AddResourceModel("files", "file", 0, true, true)
+	rm.SetValidateFormat(true)
+	rm.SetValidateCompatibility(true)
+
+	XHTTP(t, reg, "PUT", "/modelsource", model.MustUserMarshal("", "  "),
+		200, `{
+  "groups": {
+    "dirs": {
+      "plural": "dirs",
+      "singular": "dir",
+      "resources": {
+        "files": {
+          "plural": "files",
+          "singular": "file",
+          "maxversions": 0,
+          "setversionid": true,
+          "hasdocument": true,
+          "versionmode": "manual",
+          "singleversionroot": false,
+          "validateformat": true,
+          "validatecompatibility": true,
+          "strictvalidation": false
+        }
+      }
+    }
+  }
+}
+`)
+
+	// Create the target with an "unknown" format - EnsureCompat() doesn't
+	// reject this outright (only "strict" mode would), it just flags
+	// formatvalidated=false/compatibilityvalidated=false with a
+	// "Unknown format" reason.
+	XHTTP(t, reg, "PUT", "/dirs/d1/files/f1$details", `{
+  "meta": { "compatibility": "backward" },
+  "format": "unknown",
+  "file": "1"
+}`, 201, `{
+  "fileid": "f1",
+  "versionid": "1",
+  "self": "http://localhost:8181/dirs/d1/files/f1$details",
+  "xid": "/dirs/d1/files/f1",
+  "epoch": 1,
+  "isdefault": true,
+  "createdat": "2026-07-27T01:18:59.386435643Z",
+  "modifiedat": "2026-07-27T01:18:59.386435643Z",
+  "ancestorid": "1",
+  "contenttype": "application/json",
+  "format": "unknown",
+  "formatvalidated": false,
+  "formatvalidatedreason": "Unknown format",
+  "compatibilityvalidated": false,
+  "compatibilityvalidatedreason": "Unknown format",
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/f1/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/f1/versions",
+  "versionscount": 1
+}
+`)
+
+	// Create the xref source pointing at the already-failed target.
+	XHTTP(t, reg, "PUT", "/dirs/d1/files/fx/meta",
+		`{"xref":"/dirs/d1/files/f1"}`, 201, `{
+  "fileid": "fx",
+  "self": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "xid": "/dirs/d1/files/fx/meta",
+  "xref": "/dirs/d1/files/f1",
+  "epoch": 1,
+  "createdat": "2026-07-27T01:19:01.83746876Z",
+  "modifiedat": "2026-07-27T01:19:01.83746876Z",
+  "readonly": false,
+  "compatibility": "backward",
+
+  "defaultversionid": "1",
+  "defaultversionurl": "http://localhost:8181/dirs/d1/files/fx/versions/1$details",
+  "defaultversionsticky": false
+}
+`)
+
+	// The mirror must show the SAME failed state, not silently omit it
+	// or default it back to "true"/absent.
+	XHTTP(t, reg, "GET", "/dirs/d1/files/fx$details", ``, 200, `{
+  "fileid": "fx",
+  "versionid": "1",
+  "self": "http://localhost:8181/dirs/d1/files/fx$details",
+  "xid": "/dirs/d1/files/fx",
+  "epoch": 1,
+  "isdefault": true,
+  "createdat": "2026-07-27T01:19:04.344048762Z",
+  "modifiedat": "2026-07-27T01:19:04.344048762Z",
+  "ancestorid": "1",
+  "contenttype": "application/json",
+  "format": "unknown",
+  "formatvalidated": false,
+  "formatvalidatedreason": "Unknown format",
+  "compatibilityvalidated": false,
+  "compatibilityvalidatedreason": "Unknown format",
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/fx/versions",
+  "versionscount": 1
+}
+`)
+
+	// Now fix the target's format - the failure state must clear on
+	// both the target AND (more importantly) the mirror.
+	XHTTP(t, reg, "PUT", "/dirs/d1/files/f1$details", `{
+  "meta": { "compatibility": "backward" },
+  "format": "numbers",
+  "file": "1"
+}`, 200, `{
+  "fileid": "f1",
+  "versionid": "1",
+  "self": "http://localhost:8181/dirs/d1/files/f1$details",
+  "xid": "/dirs/d1/files/f1",
+  "epoch": 2,
+  "isdefault": true,
+  "createdat": "2026-07-27T01:19:06.908004939Z",
+  "modifiedat": "2026-07-27T01:19:07.055228045Z",
+  "ancestorid": "1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+  "compatibilityvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/f1/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/f1/versions",
+  "versionscount": 1
+}
+`)
+
+	XHTTP(t, reg, "GET", "/dirs/d1/files/fx$details", ``, 200, `{
+  "fileid": "fx",
+  "versionid": "1",
+  "self": "http://localhost:8181/dirs/d1/files/fx$details",
+  "xid": "/dirs/d1/files/fx",
+  "epoch": 2,
+  "isdefault": true,
+  "createdat": "2026-07-27T01:19:09.446274109Z",
+  "modifiedat": "2026-07-27T01:19:09.587498469Z",
+  "ancestorid": "1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+  "compatibilityvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/fx/versions",
+  "versionscount": 1
+}
+`)
+}
+
+// TestXrefFormatValidationFailureCascadeAfterUpdate is the "after the
+// fact" variant of TestXrefFormatValidationFailureCascade: the xref is
+// created while the target is VALID (so the mirror initially shows
+// formatvalidated=true), and only THEN does the target transition into
+// the failed ("unknown format") state via a normal direct update - the
+// mirror must pick up that failure transition too, not just the initial
+// state at xref-creation time.
+func TestXrefFormatValidationFailureCascadeAfterUpdate(t *testing.T) {
+	reg := NewRegistry("TestXrefFormatValidationFailureCascadeAfterUpdate")
+	defer PassDeleteReg(t, reg)
+
+	model := registry.Model{}
+	gm, _ := model.AddGroupModel("dirs", "dir")
+	rm, _ := gm.AddResourceModel("files", "file", 0, true, true)
+	rm.SetValidateFormat(true)
+	rm.SetValidateCompatibility(true)
+
+	XHTTP(t, reg, "PUT", "/modelsource", model.MustUserMarshal("", "  "),
+		200, `{
+  "groups": {
+    "dirs": {
+      "plural": "dirs",
+      "singular": "dir",
+      "resources": {
+        "files": {
+          "plural": "files",
+          "singular": "file",
+          "maxversions": 0,
+          "setversionid": true,
+          "hasdocument": true,
+          "versionmode": "manual",
+          "singleversionroot": false,
+          "validateformat": true,
+          "validatecompatibility": true,
+          "strictvalidation": false
+        }
+      }
+    }
+  }
+}
+`)
+
+	XHTTP(t, reg, "PUT", "/dirs/d1/files/f1$details", `{
+  "meta": { "compatibility": "backward" },
+  "format": "numbers",
+  "file": "1"
+}`, 201, `{
+  "fileid": "f1",
+  "versionid": "1",
+  "self": "http://localhost:8181/dirs/d1/files/f1$details",
+  "xid": "/dirs/d1/files/f1",
+  "epoch": 1,
+  "isdefault": true,
+  "createdat": "2026-07-27T01:19:54.525985019Z",
+  "modifiedat": "2026-07-27T01:19:54.525985019Z",
+  "ancestorid": "1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+  "compatibilityvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/f1/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/f1/versions",
+  "versionscount": 1
+}
+`)
+
+	XHTTP(t, reg, "PUT", "/dirs/d1/files/fx/meta",
+		`{"xref":"/dirs/d1/files/f1"}`, 201, `{
+  "fileid": "fx",
+  "self": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "xid": "/dirs/d1/files/fx/meta",
+  "xref": "/dirs/d1/files/f1",
+  "epoch": 1,
+  "createdat": "2026-07-27T01:19:57.177137209Z",
+  "modifiedat": "2026-07-27T01:19:57.177137209Z",
+  "readonly": false,
+  "compatibility": "backward",
+
+  "defaultversionid": "1",
+  "defaultversionurl": "http://localhost:8181/dirs/d1/files/fx/versions/1$details",
+  "defaultversionsticky": false
+}
+`)
+
+	// Confirm the mirror starts out fully valid.
+	XHTTP(t, reg, "GET", "/dirs/d1/files/fx$details", ``, 200, `{
+  "fileid": "fx",
+  "versionid": "1",
+  "self": "http://localhost:8181/dirs/d1/files/fx$details",
+  "xid": "/dirs/d1/files/fx",
+  "epoch": 1,
+  "isdefault": true,
+  "createdat": "2026-07-27T01:19:59.93952847Z",
+  "modifiedat": "2026-07-27T01:19:59.93952847Z",
+  "ancestorid": "1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+  "compatibilityvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/fx/versions",
+  "versionscount": 1
+}
+`)
+
+	// Flip the TARGET (already xref'd) to an unknown format via a
+	// normal direct update - no model change involved.
+	XHTTP(t, reg, "PUT", "/dirs/d1/files/f1$details", `{
+  "meta": { "compatibility": "backward" },
+  "format": "unknown",
+  "file": "1"
+}`, 200, `{
+  "fileid": "f1",
+  "versionid": "1",
+  "self": "http://localhost:8181/dirs/d1/files/f1$details",
+  "xid": "/dirs/d1/files/f1",
+  "epoch": 2,
+  "isdefault": true,
+  "createdat": "2026-07-27T01:20:02.787321133Z",
+  "modifiedat": "2026-07-27T01:20:02.984856352Z",
+  "ancestorid": "1",
+  "contenttype": "application/json",
+  "format": "unknown",
+  "formatvalidated": false,
+  "formatvalidatedreason": "Unknown format",
+  "compatibilityvalidated": false,
+  "compatibilityvalidatedreason": "Unknown format",
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/f1/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/f1/versions",
+  "versionscount": 1
+}
+`)
+
+	// The mirror must now reflect the failure too.
+	XHTTP(t, reg, "GET", "/dirs/d1/files/fx$details", ``, 200, `{
+  "fileid": "fx",
+  "versionid": "1",
+  "self": "http://localhost:8181/dirs/d1/files/fx$details",
+  "xid": "/dirs/d1/files/fx",
+  "epoch": 2,
+  "isdefault": true,
+  "createdat": "2026-07-27T01:20:05.655873226Z",
+  "modifiedat": "2026-07-27T01:20:05.845462478Z",
+  "ancestorid": "1",
+  "contenttype": "application/json",
+  "format": "unknown",
+  "formatvalidated": false,
+  "formatvalidatedreason": "Unknown format",
+  "compatibilityvalidated": false,
+  "compatibilityvalidatedreason": "Unknown format",
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/fx/versions",
+  "versionscount": 1
+}
+`)
+}
+
+// TestXrefTargetUpdateRejectionDoesNotCorruptMirror verifies that when a
+// direct update to the xref TARGET is REJECTED outright (a real
+// "compatibility_violation" 400, not just a formatvalidated/
+// compatibilityvalidated=false flag - see FormatNumbers.IsCompatible())
+// the target's data is left untouched AND the xref SOURCE's mirror
+// remains correctly in-sync with the target's last-good state - i.e. a
+// failed write to the target must not leave the mirror stale, corrupted,
+// or partially updated.
+func TestXrefTargetUpdateRejectionDoesNotCorruptMirror(t *testing.T) {
+	reg := NewRegistry("TestXrefTargetUpdateRejectionDoesNotCorruptMirror")
+	defer PassDeleteReg(t, reg)
+
+	model := registry.Model{}
+	gm, _ := model.AddGroupModel("dirs", "dir")
+	rm, _ := gm.AddResourceModel("files", "file", 0, true, true)
+	rm.SetValidateFormat(true)
+	rm.SetValidateCompatibility(true)
+
+	XHTTP(t, reg, "PUT", "/modelsource", model.MustUserMarshal("", "  "),
+		200, `{
+  "groups": {
+    "dirs": {
+      "plural": "dirs",
+      "singular": "dir",
+      "resources": {
+        "files": {
+          "plural": "files",
+          "singular": "file",
+          "maxversions": 0,
+          "setversionid": true,
+          "hasdocument": true,
+          "versionmode": "manual",
+          "singleversionroot": false,
+          "validateformat": true,
+          "validatecompatibility": true,
+          "strictvalidation": false
+        }
+      }
+    }
+  }
+}
+`)
+
+	// Target starts with a single Version, sum=5.
+	XHTTP(t, reg, "PUT", "/dirs/d1/files/f1$details", `{
+  "meta": { "compatibility": "backward" },
+  "format": "numbers",
+  "file": "5"
+}`, 201, `{
+  "fileid": "f1",
+  "versionid": "1",
+  "self": "http://localhost:8181/dirs/d1/files/f1$details",
+  "xid": "/dirs/d1/files/f1",
+  "epoch": 1,
+  "isdefault": true,
+  "createdat": "2026-07-27T01:20:20.723551092Z",
+  "modifiedat": "2026-07-27T01:20:20.723551092Z",
+  "ancestorid": "1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+  "compatibilityvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/f1/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/f1/versions",
+  "versionscount": 1
+}
+`)
+
+	XHTTP(t, reg, "PUT", "/dirs/d1/files/fx/meta",
+		`{"xref":"/dirs/d1/files/f1"}`, 201, `{
+  "fileid": "fx",
+  "self": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "xid": "/dirs/d1/files/fx/meta",
+  "xref": "/dirs/d1/files/f1",
+  "epoch": 1,
+  "createdat": "2026-07-27T01:20:22.72379931Z",
+  "modifiedat": "2026-07-27T01:20:22.72379931Z",
+  "readonly": false,
+  "compatibility": "backward",
+
+  "defaultversionid": "1",
+  "defaultversionurl": "http://localhost:8181/dirs/d1/files/fx/versions/1$details",
+  "defaultversionsticky": false
+}
+`)
+
+	// Confirm the mirror's initial good state.
+	XHTTP(t, reg, "GET", "/dirs/d1/files/fx$details", ``, 200, `{
+  "fileid": "fx",
+  "versionid": "1",
+  "self": "http://localhost:8181/dirs/d1/files/fx$details",
+  "xid": "/dirs/d1/files/fx",
+  "epoch": 1,
+  "isdefault": true,
+  "createdat": "2026-07-27T01:20:24.854766301Z",
+  "modifiedat": "2026-07-27T01:20:24.854766301Z",
+  "ancestorid": "1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+  "compatibilityvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/fx/versions",
+  "versionscount": 1
+}
+`)
+
+	// Attempt to add a new default Version with sum=2 (< 5) - "backward"
+	// compatibility requires the new Version's sum to be >= the old
+	// one, so this must be REJECTED outright (not just flagged false).
+	XHTTP(t, reg, "POST", "/dirs/d1/files/f1/versions", `{
+  "v2": {
+    "format": "numbers",
+    "file": "2"
+  }
+}`, 400, `{
+  "type": "https://github.com/xregistry/spec/blob/main/core/spec.md#compatibility_violation",
+  "title": "The request would cause one or more Versions of \"/dirs/d1/files/f1\" to violate its compatibility rule (backward).",
+  "detail": "Version \"/dirs/d1/files/f1/versions/v2\" (sum: 2) isn't \"backward\" compatible with \"/dirs/d1/files/f1/versions/1\" (sum: 5).",
+  "subject": "/dirs/d1/files/f1",
+  "args": {
+    "compat": "backward"
+  },
+  "source": ":registry:format_numbers:109"
+}
+`)
+
+	// The target itself must be completely unaffected by the rejected
+	// request - still just the original Version.
+	XHTTP(t, reg, "GET", "/dirs/d1/files/f1$details", ``, 200, `{
+  "fileid": "f1",
+  "versionid": "1",
+  "self": "http://localhost:8181/dirs/d1/files/f1$details",
+  "xid": "/dirs/d1/files/f1",
+  "epoch": 1,
+  "isdefault": true,
+  "createdat": "2026-07-27T01:20:29.063641818Z",
+  "modifiedat": "2026-07-27T01:20:29.063641818Z",
+  "ancestorid": "1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+  "compatibilityvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/f1/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/f1/versions",
+  "versionscount": 1
+}
+`)
+
+	// And the mirror must ALSO still be showing that same untouched,
+	// last-good state - not corrupted, not partially applied, and not
+	// pointing at a Version that was never actually created.
+	XHTTP(t, reg, "GET", "/dirs/d1/files/fx$details", ``, 200, `{
+  "fileid": "fx",
+  "versionid": "1",
+  "self": "http://localhost:8181/dirs/d1/files/fx$details",
+  "xid": "/dirs/d1/files/fx",
+  "epoch": 1,
+  "isdefault": true,
+  "createdat": "2026-07-27T01:20:31.1941946Z",
+  "modifiedat": "2026-07-27T01:20:31.1941946Z",
+  "ancestorid": "1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+  "compatibilityvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/fx/versions",
+  "versionscount": 1
+}
+`)
+}
+
+// ---- Moved from xref_model_revalidation_test.go ----
+
+func TestXrefModelRevalidationFormatCascade(t *testing.T) {
+	reg := NewRegistry("TestXrefModelRevalidationFormatCascade")
+	defer PassDeleteReg(t, reg)
+
+	gm, _ := reg.Model.AddGroupModel("dirs", "dir")
+	rm, _ := gm.AddResourceModel("files", "file", 0, true, true)
+
+	// Create target f1 with format set, but validateformat=false so
+	// formatvalidated is NOT computed/set yet.
+	XHTTP(t, reg, "PUT", "/dirs/d1/files/f1$details", `{
+  "format": "numbers",
+  "file": "1\n2\n3"
+}`, 201, `*`)
+
+	// Confirm formatvalidated is absent (validateformat is off).
+	XHTTP(t, reg, "GET", "/dirs/d1/files/f1$details", ``, 200, `*`)
+
+	// Create xref fx -> f1.
+	XHTTP(t, reg, "PUT", "/dirs/d1/files/fx/meta",
+		`{"xref":"/dirs/d1/files/f1"}`, 201, `*`)
+
+	// Confirm the xref mirror also has no formatvalidated yet.
+	XHTTP(t, reg, "GET", "/dirs/d1/files/fx$details", ``, 200, `*`)
+
+	// Now flip validateformat=true at the MODEL level (no change to f1's
+	// own data at all) - this forces Registry.VerifyData() to
+	// re-validate every existing Resource, including f1 (setting
+	// formatvalidated=true, since "1\n2\n3" is valid "numbers" format)
+	// and fx (an xref source, whose runCascade() re-run should pick up
+	// the target's fresh formatvalidated=true value).
+	rm.SetValidateFormat(true)
+	modelSrc := reg.Model.MustUserMarshal("", "  ")
+	XHTTP(t, reg, "PUT", "/modelsource", modelSrc, 200, `*`)
+
+	// The TARGET f1 must now show formatvalidated=true.
+	XHTTP(t, reg, "GET", "/dirs/d1/files/f1$details", ``, 200,
+		`^(?s)^.*"epoch": 1,.*"formatvalidated": true`)
+
+	// The xref SOURCE fx must mirror that same formatvalidated=true -
+	// this is the part that was missing coverage.
+	XHTTP(t, reg, "GET", "/dirs/d1/files/fx$details", ``, 200,
+		`^(?s)^.*"epoch": 1,.*"formatvalidated": true`)
+}
+
+// TestXrefModelRevalidationCompatCascade is the compatibilityvalidated
+// analog of TestXrefModelRevalidationFormatCascade above (gaps item 1's
+// remaining "compatibilityvalidated variant" sub-case) - it checks that
+// when a model change forces Registry.VerifyData() to re-validate an
+// xref TARGET's compatibilityvalidated system attribute (with no other
+// attribute change), that fresh value is correctly cascaded into any
+// xref SOURCE pointing at it, exactly as formatvalidated is above.
+func TestXrefModelRevalidationCompatCascade(t *testing.T) {
+	reg := NewRegistry("TestXrefModelRevalidationCompatCascade")
+	defer PassDeleteReg(t, reg)
+
+	gm, _ := reg.Model.AddGroupModel("dirs", "dir")
+	rm, _ := gm.AddResourceModel("files", "file", 0, true, true)
+	rm.SetValidateFormat(true)
+
+	// Create target f1 with two versions (equal sums, so compatible
+	// under any mode), but validatecompatibility=false so
+	// compatibilityvalidated is NOT computed/set yet.
+	XHTTP(t, reg, "PUT", "/dirs/d1/files/f1$details", `{
+  "meta": { "compatibility": "backward" },
+  "versions": {
+    "v1": { "format": "numbers", "file": "2" },
+    "v2": { "format": "numbers", "file": "2" }
+  }
+}`, 201, `{
+  "fileid": "f1",
+  "versionid": "v2",
+  "self": "http://localhost:8181/dirs/d1/files/f1$details",
+  "xid": "/dirs/d1/files/f1",
+  "epoch": 1,
+  "isdefault": true,
+  "createdat": "2026-07-27T00:28:13.175487333Z",
+  "modifiedat": "2026-07-27T00:28:13.175487333Z",
+  "ancestorid": "v1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/f1/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/f1/versions",
+  "versionscount": 2
+}
+`)
+
+	// Confirm compatibilityvalidated is absent (validatecompatibility is
+	// off).
+	XHTTP(t, reg, "GET", "/dirs/d1/files/f1$details", ``, 200, `{
+  "fileid": "f1",
+  "versionid": "v2",
+  "self": "http://localhost:8181/dirs/d1/files/f1$details",
+  "xid": "/dirs/d1/files/f1",
+  "epoch": 1,
+  "isdefault": true,
+  "createdat": "2026-07-27T00:28:15.548850155Z",
+  "modifiedat": "2026-07-27T00:28:15.548850155Z",
+  "ancestorid": "v1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/f1/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/f1/versions",
+  "versionscount": 2
+}
+`)
+
+	// Create xref fx -> f1.
+	XHTTP(t, reg, "PUT", "/dirs/d1/files/fx/meta",
+		`{"xref":"/dirs/d1/files/f1"}`, 201, `{
+  "fileid": "fx",
+  "self": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "xid": "/dirs/d1/files/fx/meta",
+  "xref": "/dirs/d1/files/f1",
+  "epoch": 1,
+  "createdat": "2026-07-27T00:28:17.538190459Z",
+  "modifiedat": "2026-07-27T00:28:17.538190459Z",
+  "readonly": false,
+  "compatibility": "backward",
+
+  "defaultversionid": "v2",
+  "defaultversionurl": "http://localhost:8181/dirs/d1/files/fx/versions/v2$details",
+  "defaultversionsticky": false
+}
+`)
+
+	// Confirm the xref mirror also has no compatibilityvalidated yet.
+	XHTTP(t, reg, "GET", "/dirs/d1/files/fx$details", ``, 200, `{
+  "fileid": "fx",
+  "versionid": "v2",
+  "self": "http://localhost:8181/dirs/d1/files/fx$details",
+  "xid": "/dirs/d1/files/fx",
+  "epoch": 1,
+  "isdefault": true,
+  "createdat": "2026-07-27T00:28:19.619210932Z",
+  "modifiedat": "2026-07-27T00:28:19.619210932Z",
+  "ancestorid": "v1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/fx/versions",
+  "versionscount": 2
+}
+`)
+
+	// Now flip validatecompatibility=true at the MODEL level (no change
+	// to f1's own data at all) - this forces Registry.VerifyData() to
+	// re-validate every existing Resource, including f1 (setting
+	// compatibilityvalidated=true, since v1/v2 are compatible under
+	// "backward") and fx (an xref source, whose runCascade() re-run
+	// should pick up the target's fresh compatibilityvalidated=true
+	// value).
+	rm.SetValidateCompatibility(true)
+	modelSrc := reg.Model.MustUserMarshal("", "  ")
+	XHTTP(t, reg, "PUT", "/modelsource", modelSrc, 200, `{
+  "attributes": {
+    "specversion": {
+      "name": "specversion",
+      "type": "string",
+      "readonly": true,
+      "required": true
+    },
+    "registryid": {
+      "name": "registryid",
+      "type": "string",
+      "matchcase": true,
+      "readonly": true,
+      "immutable": true,
+      "required": true
+    },
+    "self": {
+      "name": "self",
+      "type": "url",
+      "readonly": true,
+      "immutable": true,
+      "required": true
+    },
+    "shortself": {
+      "name": "shortself",
+      "type": "url",
+      "readonly": true,
+      "immutable": true
+    },
+    "xid": {
+      "name": "xid",
+      "type": "xid",
+      "readonly": true,
+      "immutable": true,
+      "required": true
+    },
+    "epoch": {
+      "name": "epoch",
+      "type": "uinteger",
+      "readonly": true,
+      "required": true
+    },
+    "name": {
+      "name": "name",
+      "type": "string"
+    },
+    "description": {
+      "name": "description",
+      "type": "string"
+    },
+    "documentation": {
+      "name": "documentation",
+      "type": "url"
+    },
+    "icon": {
+      "name": "icon",
+      "type": "url"
+    },
+    "labels": {
+      "name": "labels",
+      "type": "map",
+      "item": {
+        "type": "string"
+      }
+    },
+    "createdat": {
+      "name": "createdat",
+      "type": "timestamp",
+      "required": true
+    },
+    "modifiedat": {
+      "name": "modifiedat",
+      "type": "timestamp",
+      "required": true
+    },
+    "capabilities": {
+      "name": "capabilities",
+      "type": "object",
+      "attributes": {
+        "*": {
+          "name": "*",
+          "type": "any"
+        }
+      }
+    },
+    "model": {
+      "name": "model",
+      "type": "object",
+      "readonly": true,
+      "attributes": {
+        "*": {
+          "name": "*",
+          "type": "any"
+        }
+      }
+    },
+    "modelsource": {
+      "name": "modelsource",
+      "type": "object",
+      "attributes": {
+        "*": {
+          "name": "*",
+          "type": "any"
+        }
+      }
+    },
+    "dirsurl": {
+      "name": "dirsurl",
+      "type": "url",
+      "readonly": true,
+      "immutable": true,
+      "required": true
+    },
+    "dirscount": {
+      "name": "dirscount",
+      "type": "uinteger",
+      "readonly": true,
+      "required": true
+    },
+    "dirs": {
+      "name": "dirs",
+      "type": "map",
+      "item": {
+        "type": "object",
+        "attributes": {
+          "*": {
+            "name": "*",
+            "type": "any"
+          }
+        }
+      }
+    }
+  },
+  "groups": {
+    "dirs": {
+      "plural": "dirs",
+      "singular": "dir",
+      "attributes": {
+        "dirid": {
+          "name": "dirid",
+          "type": "string",
+          "matchcase": true,
+          "immutable": true,
+          "required": true
+        },
+        "self": {
+          "name": "self",
+          "type": "url",
+          "readonly": true,
+          "immutable": true,
+          "required": true
+        },
+        "shortself": {
+          "name": "shortself",
+          "type": "url",
+          "readonly": true,
+          "immutable": true
+        },
+        "xid": {
+          "name": "xid",
+          "type": "xid",
+          "readonly": true,
+          "immutable": true,
+          "required": true
+        },
+        "epoch": {
+          "name": "epoch",
+          "type": "uinteger",
+          "readonly": true,
+          "required": true
+        },
+        "name": {
+          "name": "name",
+          "type": "string"
+        },
+        "description": {
+          "name": "description",
+          "type": "string"
+        },
+        "documentation": {
+          "name": "documentation",
+          "type": "url"
+        },
+        "icon": {
+          "name": "icon",
+          "type": "url"
+        },
+        "labels": {
+          "name": "labels",
+          "type": "map",
+          "item": {
+            "type": "string"
+          }
+        },
+        "createdat": {
+          "name": "createdat",
+          "type": "timestamp",
+          "required": true
+        },
+        "modifiedat": {
+          "name": "modifiedat",
+          "type": "timestamp",
+          "required": true
+        },
+        "deprecated": {
+          "name": "deprecated",
+          "type": "object",
+          "attributes": {
+            "alternative": {
+              "name": "alternative",
+              "type": "url"
+            },
+            "documentation": {
+              "name": "documentation",
+              "type": "url"
+            },
+            "effective": {
+              "name": "effective",
+              "type": "timestamp"
+            },
+            "removal": {
+              "name": "removal",
+              "type": "timestamp"
+            },
+            "*": {
+              "name": "*",
+              "type": "any"
+            }
+          }
+        },
+        "constraints": {
+          "name": "constraints",
+          "type": "map",
+          "item": {
+            "type": "object",
+            "attributes": {
+              "default": {
+                "name": "default",
+                "type": "any"
+              },
+              "enum": {
+                "name": "enum",
+                "type": "array",
+                "item": {
+                  "type": "any"
+                }
+              },
+              "equals": {
+                "name": "equals",
+                "type": "string"
+              }
+            }
+          }
+        },
+        "filesurl": {
+          "name": "filesurl",
+          "type": "url",
+          "readonly": true,
+          "immutable": true,
+          "required": true
+        },
+        "filescount": {
+          "name": "filescount",
+          "type": "uinteger",
+          "readonly": true,
+          "required": true
+        },
+        "files": {
+          "name": "files",
+          "type": "map",
+          "item": {
+            "type": "object",
+            "attributes": {
+              "*": {
+                "name": "*",
+                "type": "any"
+              }
+            }
+          }
+        }
+      },
+      "resources": {
+        "files": {
+          "plural": "files",
+          "singular": "file",
+          "maxversions": 0,
+          "setversionid": true,
+          "hasdocument": true,
+          "versionmode": "manual",
+          "singleversionroot": false,
+          "validateformat": true,
+          "validatecompatibility": true,
+          "strictvalidation": false,
+          "attributes": {
+            "fileid": {
+              "name": "fileid",
+              "type": "string",
+              "matchcase": true,
+              "immutable": true,
+              "required": true
+            },
+            "versionid": {
+              "name": "versionid",
+              "type": "string",
+              "matchcase": true,
+              "immutable": true,
+              "required": true
+            },
+            "self": {
+              "name": "self",
+              "type": "url",
+              "readonly": true,
+              "immutable": true,
+              "required": true
+            },
+            "shortself": {
+              "name": "shortself",
+              "type": "url",
+              "readonly": true,
+              "immutable": true
+            },
+            "xid": {
+              "name": "xid",
+              "type": "xid",
+              "readonly": true,
+              "immutable": true,
+              "required": true
+            },
+            "epoch": {
+              "name": "epoch",
+              "type": "uinteger",
+              "readonly": true,
+              "required": true
+            },
+            "name": {
+              "name": "name",
+              "type": "string"
+            },
+            "isdefault": {
+              "name": "isdefault",
+              "type": "boolean",
+              "readonly": true,
+              "required": true,
+              "default": false
+            },
+            "description": {
+              "name": "description",
+              "type": "string"
+            },
+            "documentation": {
+              "name": "documentation",
+              "type": "url"
+            },
+            "icon": {
+              "name": "icon",
+              "type": "url"
+            },
+            "labels": {
+              "name": "labels",
+              "type": "map",
+              "item": {
+                "type": "string"
+              }
+            },
+            "createdat": {
+              "name": "createdat",
+              "type": "timestamp",
+              "required": true
+            },
+            "modifiedat": {
+              "name": "modifiedat",
+              "type": "timestamp",
+              "required": true
+            },
+            "ancestorid": {
+              "name": "ancestorid",
+              "type": "string",
+              "matchcase": true,
+              "required": true
+            },
+            "contenttype": {
+              "name": "contenttype",
+              "type": "string"
+            },
+            "format": {
+              "name": "format",
+              "type": "string"
+            },
+            "formatvalidated": {
+              "name": "formatvalidated",
+              "type": "boolean",
+              "readonly": true
+            },
+            "formatvalidatedreason": {
+              "name": "formatvalidatedreason",
+              "type": "string",
+              "readonly": true
+            },
+            "compatibilityvalidated": {
+              "name": "compatibilityvalidated",
+              "type": "boolean",
+              "readonly": true
+            },
+            "compatibilityvalidatedreason": {
+              "name": "compatibilityvalidatedreason",
+              "type": "string",
+              "readonly": true
+            },
+            "fileurl": {
+              "name": "fileurl",
+              "type": "url"
+            },
+            "fileproxyurl": {
+              "name": "fileproxyurl",
+              "type": "url"
+            },
+            "file": {
+              "name": "file",
+              "type": "any"
+            },
+            "filebase64": {
+              "name": "filebase64",
+              "type": "string"
+            }
+          },
+          "resourceattributes": {
+            "fileid": {
+              "name": "fileid",
+              "type": "string",
+              "matchcase": true,
+              "immutable": true,
+              "required": true
+            },
+            "self": {
+              "name": "self",
+              "type": "url",
+              "readonly": true,
+              "immutable": true,
+              "required": true
+            },
+            "shortself": {
+              "name": "shortself",
+              "type": "url",
+              "readonly": true,
+              "immutable": true
+            },
+            "xid": {
+              "name": "xid",
+              "type": "xid",
+              "readonly": true,
+              "immutable": true,
+              "required": true
+            },
+            "metaurl": {
+              "name": "metaurl",
+              "type": "url",
+              "readonly": true,
+              "immutable": true,
+              "required": true
+            },
+            "meta": {
+              "name": "meta",
+              "type": "object",
+              "attributes": {
+                "*": {
+                  "name": "*",
+                  "type": "any"
+                }
+              }
+            },
+            "versionsurl": {
+              "name": "versionsurl",
+              "type": "url",
+              "readonly": true,
+              "immutable": true,
+              "required": true
+            },
+            "versionscount": {
+              "name": "versionscount",
+              "type": "uinteger",
+              "readonly": true,
+              "required": true
+            },
+            "versions": {
+              "name": "versions",
+              "type": "map",
+              "item": {
+                "type": "object",
+                "attributes": {
+                  "*": {
+                    "name": "*",
+                    "type": "any"
+                  }
+                }
+              }
+            }
+          },
+          "metaattributes": {
+            "fileid": {
+              "name": "fileid",
+              "type": "string",
+              "matchcase": true,
+              "immutable": true,
+              "required": true
+            },
+            "self": {
+              "name": "self",
+              "type": "url",
+              "readonly": true,
+              "immutable": true,
+              "required": true
+            },
+            "shortself": {
+              "name": "shortself",
+              "type": "url",
+              "readonly": true,
+              "immutable": true
+            },
+            "xid": {
+              "name": "xid",
+              "type": "xid",
+              "readonly": true,
+              "immutable": true,
+              "required": true
+            },
+            "xref": {
+              "name": "xref",
+              "type": "url"
+            },
+            "epoch": {
+              "name": "epoch",
+              "type": "uinteger",
+              "readonly": true,
+              "required": true
+            },
+            "labels": {
+              "name": "labels",
+              "type": "map",
+              "item": {
+                "type": "string"
+              }
+            },
+            "createdat": {
+              "name": "createdat",
+              "type": "timestamp",
+              "required": true
+            },
+            "modifiedat": {
+              "name": "modifiedat",
+              "type": "timestamp",
+              "required": true
+            },
+            "readonly": {
+              "name": "readonly",
+              "type": "boolean",
+              "readonly": true,
+              "required": true,
+              "default": false
+            },
+            "compatibility": {
+              "name": "compatibility",
+              "type": "string",
+              "enum": [
+                "backward",
+                "backward_transitive",
+                "forward",
+                "forward_transitive",
+                "full",
+                "full_transitive"
+              ],
+              "strict": true
+            },
+            "deprecated": {
+              "name": "deprecated",
+              "type": "object",
+              "attributes": {
+                "alternative": {
+                  "name": "alternative",
+                  "type": "url"
+                },
+                "documentation": {
+                  "name": "documentation",
+                  "type": "url"
+                },
+                "effective": {
+                  "name": "effective",
+                  "type": "timestamp"
+                },
+                "removal": {
+                  "name": "removal",
+                  "type": "timestamp"
+                },
+                "*": {
+                  "name": "*",
+                  "type": "any"
+                }
+              }
+            },
+            "defaultversionid": {
+              "name": "defaultversionid",
+              "type": "string",
+              "matchcase": true,
+              "required": true
+            },
+            "defaultversionurl": {
+              "name": "defaultversionurl",
+              "type": "url",
+              "readonly": true,
+              "required": true
+            },
+            "defaultversionsticky": {
+              "name": "defaultversionsticky",
+              "type": "boolean",
+              "required": true,
+              "default": false
+            }
+          }
+        }
+      }
+    }
+  }
+}
+`)
+
+	// The TARGET f1 must now show compatibilityvalidated=true.
+	XHTTP(t, reg, "GET", "/dirs/d1/files/f1$details", ``, 200,
+		`{
+  "fileid": "f1",
+  "versionid": "v2",
+  "self": "http://localhost:8181/dirs/d1/files/f1$details",
+  "xid": "/dirs/d1/files/f1",
+  "epoch": 1,
+  "isdefault": true,
+  "createdat": "2026-07-27T00:28:23.801192485Z",
+  "modifiedat": "2026-07-27T00:28:23.801192485Z",
+  "ancestorid": "v1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+  "compatibilityvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/f1/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/f1/versions",
+  "versionscount": 2
+}
+`)
+
+	// The xref SOURCE fx must mirror that same
+	// compatibilityvalidated=true - this is the part that was missing
+	// coverage.
+	XHTTP(t, reg, "GET", "/dirs/d1/files/fx$details", ``, 200,
+		`{
+  "fileid": "fx",
+  "versionid": "v2",
+  "self": "http://localhost:8181/dirs/d1/files/fx$details",
+  "xid": "/dirs/d1/files/fx",
+  "epoch": 1,
+  "isdefault": true,
+  "createdat": "2026-07-27T00:28:25.925416292Z",
+  "modifiedat": "2026-07-27T00:28:25.925416292Z",
+  "ancestorid": "v1",
+  "contenttype": "application/json",
+  "format": "numbers",
+  "formatvalidated": true,
+  "compatibilityvalidated": true,
+
+  "metaurl": "http://localhost:8181/dirs/d1/files/fx/meta",
+  "versionsurl": "http://localhost:8181/dirs/d1/files/fx/versions",
+  "versionscount": 2
+}
+`)
+}
