+++ tests/misc_test.go
@@ -55,7 +55,7 @@ func TestMiscDBRows(t *testing.T) {
 	}
 
 	rows := reg.Query("SELECT Path,PropName,PropValue "+
-		"FROM FullTreeTable WHERE RegSID=? AND IsDefaultVerCopy=false AND "+
+		"FROM Props WHERE RegSID=? AND IsDefaultVerCopy=false AND "+
 		"IsXrefPropCopy=false AND IsXrefVerCopy=false AND "+
 		"IsCalcStatic=false AND IsCalcDynamic=false "+
 		"ORDER BY Path, PropName ",
